Day4/requirements.txt
google-genai
python-dotenv
openai
anthropic
tenacity
pydantic
streamlit